+++ src/main/java/lazyae2/tile/TileAggregator.java
@@ -128,6 +128,11 @@ public final class TileAggregator extends TileProcessor {
     }
 
     @Override
+    public ItemStack getWorkOutput() {
+        return this.recipe == null ? ItemStack.EMPTY : this.recipe.getOutput();
+    }
+
+    @Override
     protected boolean recomputeCanWork() {
         final AggregatorRecipe found = LazyRecipes.findAggregator(this.slots());
         if (found == null) {